Make clippy happy again
Remaining new/default-warning in coreis a good reminder that the zero-init
might be wrong.

+++ super-rustycom-client/src/ui/context.rs
@@ -21,7 +21,7 @@ impl Context {
         let mut platform = WinitPlatform::init(&mut context);
 
         // This is where highdpi would go, but we always use physical size
-        let font_size = 13.0 as f32;
+        let font_size = 13.0;
         context.fonts().add_font(&[FontSource::DefaultFontData {
             config: Some(FontConfig {
                 size_pixels: font_size,
